+++ src/main/java/com/qring/coupon/infrastructure/docs/CouponControllerSwagger.java
@@ -1,12 +1,12 @@
 package com.qring.coupon.infrastructure.docs;
 
 import com.qring.coupon.application.global.dto.ResDTO;
-import com.qring.coupon.application.v1.res.CouponGetByIdResDTOV1;
-import com.qring.coupon.application.v1.res.CouponPostByIdResDTOV1;
-import com.qring.coupon.application.v1.res.CouponPostResDTOV1;
-import com.qring.coupon.application.v1.res.CouponSearchResDTOV1;
-import com.qring.coupon.presentation.v1.req.PostCouponReqDTOV1;
-import com.qring.coupon.presentation.v1.req.PutCouponReqDTOV1;
+import com.qring.coupon.application.res.CouponGetByIdResDTOV1;
+import com.qring.coupon.application.res.CouponPostByIdResDTOV1;
+import com.qring.coupon.application.res.CouponPostResDTOV1;
+import com.qring.coupon.application.res.CouponSearchResDTOV1;
+import com.qring.coupon.presentation.req.PostCouponReqDTOV1;
+import com.qring.coupon.presentation.req.PutCouponReqDTOV1;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
